refactor(slack-app): move posthog-code workflows and eval into slack_app
Co-locate the two posthog-code workflow files and the repo-selection
eval script with the activity package they already depend on:

- posthog/temporal/ai/posthog_code_slack_mention.py → posthog/temporal/ai/slack_app/
- posthog/temporal/ai/posthog_code_slack_mention_command.py → posthog/temporal/ai/slack_app/
- posthog/temporal/ai/eval_slack_repo_selection.py → posthog/temporal/ai/slack_app/

git mv preserves history. Updates importers in
posthog/temporal/ai/__init__.py, products/slack_app/backend/api.py,
products/slack_app/backend/tests/test_posthog_code_event_handler.py,
and the eval-script path reference in posthog/temporal/AGENTS.md.
